PublicKey tests.
